requirements.txt
appdirs==1.4.3
asgiref==3.3.1
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.4
black==19.10b0
certifi==2020.11.8
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.4.2
Django==3.1.3
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
emoji==0.6.0
future==0.18.2
idna==2.10
isort==4.3.21
joblib==1.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.2.0
mccabe==0.6.1
networkx==2.4
nltk==3.3
numpy==1.18.1
oauthlib==3.1.0
pandas==1.1.4
pathspec==0.7.0
Pillow==8.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
PySocks==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.4
regex==2020.2.20
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.13.0
sqlparse==0.4.1
tagme==0.1.3
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.0
tweepy==3.9.0
typed-ast==1.4.1
Unidecode==1.1.2
urllib3==1.25.11
vaderSentiment==3.3.2
wordcloud==1.8.0
wrapt==1.11.2
